♻️ [Refactor] : Community 부분 컴포넌트 분리
Community(리스트) 컴포넌트, CommunityDetail(상세) 컴포넌트, CommunityForm(추가/수정) 컴포넌트를
section 단위로 구분하여 컴포넌트 분리하였습니다.
이후, CSS 모듈화, 반응형 CSS를 추가하여 수정할 예정입니다.

Related to #1 #2 #3
