+++ src/components/Common/BannerSection.jsx
@@ -4,8 +4,8 @@ import { PropTypes } from 'prop-types';
 function BannerSection({ category, part }) {
   return (
     <section className="w-full py-20 flex flex-col text-center gap-8">
-      <h3 className="text-open-font-xxl">{category}</h3>
-      <h2 className="text-open-font-xxxxl font-bold">{part}</h2>
+      <h2 className="text-open-font-xxl">{category}</h2>
+      <h3 className="text-open-font-xxxxl font-bold">{part}</h3>
     </section>
   );
 }
