+++ src/components/Community/NoticeList.jsx
@@ -21,27 +21,31 @@ NoticeList.propTypes = {
   notices: PropTypes.arrayOf(
     PropTypes.shape({
       id: PropTypes.string.isRequired,
-      index: PropTypes.number.isRequired,
-      title: PropTypes.string.isRequired,
-      updatedAt: PropTypes.shape({
-        seconds: PropTypes.number.isRequired,
-        nanoseconds: PropTypes.number.isRequired,
+      data: PropTypes.shape({
+        index: PropTypes.number,
+        title: PropTypes.string,
+        content: PropTypes.string,
+        imageUrl: PropTypes.string,
+        updatedAt: PropTypes.shape({
+          seconds: PropTypes.number,
+          nanoseconds: PropTypes.number,
+        }),
       }).isRequired,
-      imageUrl: PropTypes.string.isRequired,
-      content: PropTypes.string.isRequired,
     }),
   ).isRequired,
   currentNotices: PropTypes.arrayOf(
     PropTypes.shape({
       id: PropTypes.string.isRequired,
-      index: PropTypes.number.isRequired,
-      title: PropTypes.string.isRequired,
-      updatedAt: PropTypes.shape({
-        seconds: PropTypes.number.isRequired,
-        nanoseconds: PropTypes.number.isRequired,
+      data: PropTypes.shape({
+        index: PropTypes.number,
+        title: PropTypes.string,
+        content: PropTypes.string,
+        imageUrl: PropTypes.string,
+        updatedAt: PropTypes.shape({
+          seconds: PropTypes.number,
+          nanoseconds: PropTypes.number,
+        }),
       }).isRequired,
-      imageUrl: PropTypes.string.isRequired,
-      content: PropTypes.string.isRequired,
     }),
   ).isRequired,
 };
